revert `Do floating point math` because i don't want static cast all the returns in the function. better approach would be divide this function in two, one for raw value that return int, one for percent value that return double or function that return healthy kcal directly

+++ src/math_parser_diag.cpp
@@ -1338,7 +1338,7 @@ std::function<double( dialogue & )> calories_eval( char scope,
 
         if( format == "percent" ) {
             if( d.actor( beta )->get_character() ) {
-                double divisor = d.actor( beta )->get_healthy_kcal() / 100.0;
+                int divisor = d.actor( beta )->get_healthy_kcal() / 100.0;
                 //if no data, default to default height of 175cm
                 if( divisor == 0 ) {
                     debugmsg( "Can't get healthy amount of calories, return raw calories instead" );
